% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the StatFlow Shiny Application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[golem]{with_golem_options}}.}
}
\value{
No return value, called for side effects.
}
\description{
Launches the StatFlow interactive introductory data analysis
platform in your default web browser.
}
\examples{
\dontrun{
  StatFlow::run_app()
}

}
